requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlmodel==0.0.14
python-multipart==0.0.7
passlib[bcrypt]==1.7.4
python-jose==3.3.0
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
